COM-3202: move image
